backend/requirements.txt
# FastAPI and ASGI server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0

# DocETL for document analysis pipeline
docetl>=0.1.0

# Environment variables
python-dotenv>=1.0.0

# PDF generation
reportlab>=4.2.0
